feat(backend): Add Cloudinary Configuration
Adds the required configuration for Cloudinary in the application properties file. This includes the cloudinary name, API key and API secret.

A configuration class for Cloudinary is also added, which creates a Cloudinary bean with the provided configuration values.

A Cloudinary service is added that allows uploading images and generating image URLs with Cloudinary.