feat(backend): Add endpoint to update user profile
Added a new PATCH endpoint in the UserController controller to allow users to update their profile. Implemented updateProfile method that receives a profile update request and user authentication. Updated the user's first name, last name, username, email and date of birth fields in the database. In addition, the functionality to upload an avatar image and save its URL to the user's profile was added.