feat(roadmap): Add user context to getRoadmapBySlug (#87)
The changes in this commit add the user ID as a parameter to the `getRoadmapBySlug` method in the `RoadmapService`. This allows the service to retrieve the roadmap details with the user's progress information, such as which exercises have been completed. This feature is necessary to provide a personalized view of the roadmap for each user.