feat: Add display order to entities (#95)
* feat: improve error handling and logging

This commit introduces the following changes:

- Import the `error` function from the `@tauri-apps/plugin-log` module to log errors more effectively.
- Update the `showSystemNotification` function to display a more user-
  friendly error message, including the error status code.
- Replace the `throw` statement with a call to the `error` function to
  log the fatal error details, including the error message and stack trace.
- Remove the commented-out `throw` statement, as the error is now being
  logged instead of thrown.

These changes aim to improve the error handling and logging in the
desktop application, providing more informative and user-friendly error
messages, as well as better logging of fatal errors for debugging
purposes.

* feat(education): Add display_order field to Module, Lesson, and Exercise

This commit adds a new `display_order` field to the `ModuleResponse`, `CourseUpdateRequest`, `Exercise`, `RoadmapExerciseDTO`, `LessonResponse`, `RoadmapLessonDTO`, `ModuleRequest`, and `ModuleUpdateRequest` classes. This field will be used to store the order in which the modules, lessons, and exercises should be displayed to the user.

The `display_order` field is added to the `ModuleEntity` and `Exercise` entities, and an index is added to the `ModuleEntity` table to allow for efficient sorting and retrieval of modules based on their display order.

This change will enable the application to display the modules, lessons, and exercises in a specific order, improving the user experience and making it easier for users to navigate the educational content.

* feat(exercise): Add display_order field to exercise-related DTOs

The changes add a new `display_order` field to various exercise-related DTOs, including `ExerciseResponse`, `ExerciseConfig`, `UpdateExercise`, `CreateExercise`, and `ExerciseDetailsResponse`. This field will be used to control the order in which exercises are displayed to users.