update-dependencies (#107)
* feat(roadmap): Improve drag and drop handling in RoadmapForm

Adds event handlers to prevent the default browser behavior when
dragging and dropping files onto the RoadmapForm component. This
ensures a better user experience by allowing the component to handle
the file drop event properly.

Additionally, moves the import statements for the `AppError`,
`RoadmapCreateRequest`, and `TagDTO` types to the top of the file to
improve code organization and readability.

* feat(deps): update thiserror and tokio dependencies

Update the thiserror and tokio dependencies to their latest versions.
The thiserror dependency is updated from 2.0.0 to 2.0.1 to get the latest
bug fixes and improvements. The tokio dependency is updated from 1.41.0
to 1.41.1 to get the latest bug fixes and improvements as well.

* feat(pnpm-lock.yaml): update dependencies

- Update `radix-vue` to version `1.9.9`
- Update `vee-validate` to version `4.14.7`
- Update `lucide-vue-next` to version `0.456.0`
- Update `@nuxt/kit` to version `3.14.159`
- Update `@vue/devtools-kit` to version `7.6.3`

These updates ensure the latest versions of the dependencies are used, which may include bug fixes, performance improvements, or new features.