[WIP] Splashscreen design (#111)
* feat(roadmap): Improve roadmap form UI and remove unused features

The changes in this commit focus on improving the user interface of the roadmap form and removing unused features. The key changes are:

- Simplified the roadmap form by removing the "Courses" section, as it was not being used.
- Improved the layout and styling of the "Tags" section, making it more visually appealing.
- Optimized the code by removing unused variables and imports.
- Improved the formatting and consistency of the code.

These changes aim to enhance the user experience and streamline the roadmap creation process.

* feat(splashscreen): Implement new splash screen design

The changes in this commit focus on the implementation of a new splash screen design for the desktop application. The key changes are:

- Replaced the previous simple splash screen with a more visually appealing and interactive design.
- Added a left column with the app logo, loading text, and links to social media channels.
- Implemented a right column with an illustration and decorative elements to create a more engaging user experience.
- Configured the splash screen window in the Tauri configuration file to be centered, non-resizable, and always on top.
- Added new image assets for the splash screen illustration and logo.

These changes aim to provide a more polished and visually appealing splash screen experience for the users of the desktop application.